requirements.txt
flask
flask_cors
langchain
langchain_community
langchain_groq
langchain-text-splitters
faiss-cpu
sentence-transformers
python-dotenv